+++ src/Controller/VilleController.php
@@ -45,19 +45,26 @@ class VilleController extends AbstractController
      * @Route("/ville", name="ville")
      */
     public function ville(Request $request)
-    { //recuperation du repository grace au manager
-        $em = $this->getDoctrine()->getManager();
-        $villeRepository = $em->getRepository(Ville::class);
-        //personneRepository herite de servciceEntityRepository ayant les methodes pour recuperer les données de la bdd
-        $listeVilles = $villeRepository->findAll();
-        $resultat = [];
-        foreach ($listeVilles as $vil) {
-            array_push($resultat, ["id"=>$vil->getId(),"nom"=>$vil->getNom(),"codepostal"=>$vil->getCodepostal()]);
+    { 
+        if ($request->isMethod('get')) {
+
+
+            //recuperation du repository grace au manager
+            $em = $this->getDoctrine()->getManager();
+            $villeRepository = $em->getRepository(Ville::class);
+            //personneRepository herite de servciceEntityRepository ayant les methodes pour recuperer les données de la bdd
+            $listeVilles = $villeRepository->findAll();
+            $resultat = [];
+            foreach ($listeVilles as $vil) {
+                array_push($resultat, ["id"=>$vil->getId(),"nom"=>$vil->getNom(),"codepostal"=>$vil->getCodepostal()]);
+            }
+            $reponse = new JsonResponse($resultat);
+    
+    
+            return $reponse;
+        } else {
+            return new Response('Failed', 404);
         }
-        $reponse = new JsonResponse($resultat);
-
-
-        return $reponse;
     }
 
 
@@ -68,16 +75,21 @@ class VilleController extends AbstractController
 
     public function delete(Request $request, $id)
     {
-        //récupération du Manager  et du repository pour accéder à la bdd
-        $em = $this->getDoctrine()->getManager();
-        $villeRepository = $em->getRepository(Ville::class);
-        //requete de selection
-        $vil = $villeRepository->find($id);
-        //suppression de l'entity
-        $em->remove($vil);
-        $em->flush();
-        $resultat = ["ok"];
-        $reponse = new JsonResponse($resultat);
-        return $reponse;
+        if ($request->isMethod('delete')){
+
+            //récupération du Manager  et du repository pour accéder à la bdd
+            $em = $this->getDoctrine()->getManager();
+            $villeRepository = $em->getRepository(Ville::class);
+            //requete de selection
+            $vil = $villeRepository->find($id);
+            //suppression de l'entity
+            $em->remove($vil);
+            $em->flush();
+            $resultat = ["ok"];
+            $reponse = new JsonResponse($resultat);
+            return $reponse;
+        } else {
+            return new Response('Failed', 404);
+        }
     }
 }
